+++ onto_spread_ed/database/Base.py
@@ -1,0 +1,5 @@
+from sqlalchemy.ext.declarative import declarative_base
+
+Base = declarative_base()
+
+
