+++ onto_spread_ed/guards/verify_login.py
@@ -11,7 +11,7 @@ def verify_logged_in(fn):
     def wrapped(*args, **kwargs):
         # If the user is not logged in, then redirect him to the "logged out" page:
         if not g.user:
-            return redirect(url_for("login"))
+            return redirect(url_for("authentication.login"))
         return fn(*args, **kwargs)
 
     return wrapped
