+++ onto_spread_ed/index/FileStorage.py
@@ -1,6 +1,6 @@
 import whoosh.filedb.filestore
 
-from index.ExtendedStorage import ExtendedStorage
+from .ExtendedStorage import ExtendedStorage
 
 
 class FileStorage(ExtendedStorage, whoosh.filedb.filestore.FileStorage):
