+++ onto_spread_ed/routes/admin.py
@@ -1,0 +1,4 @@
+from flask import Blueprint
+
+bp = Blueprint("admin", __name__, url_prefix="admin")
+
