+++ onto_spread_ed/routes/auth.py
@@ -1,0 +1,34 @@
+from flask import Blueprint, session, redirect, url_for, render_template, jsonify, g
+from flask_github import GitHub
+
+from ..database.User import User
+from ..guards.verify_login import verify_logged_in
+
+bp = Blueprint("authentication", __name__, template_folder="../templates")
+
+@bp.route('/login')
+def login(github: GitHub):
+    if session.get('user_id', None) is not None:
+        session.pop('user_id', None)  # Could be stale
+
+    return github.authorize(scope="user,repo")
+
+
+@bp.route('/logout')
+def logout():
+    session.pop('user_id', None)
+    return redirect(url_for('authentication.loggedout'))
+
+
+@bp.route("/loggedout")
+def loggedout():
+    """
+    Displays the page to be shown to logged out users.
+    """
+    return render_template("loggedout.html")
+
+
+@bp.route('/user')
+@verify_logged_in
+def user(github: GitHub):
+    return jsonify(github.get('/user'))
