+++ onto_spread_ed/templates/testgeneratedbyapi.html
@@ -25,7 +25,7 @@
 		integrity="sha384-Vkoo8x4CGsO3+Hhxv8T/Q5PaXtkKtu6ug5TOeNV6gBiFeWPGFN9MuhOf23Q9Ifjh" crossorigin="anonymous">
 
 	<!-- Bootstrap filter table CSS -->
-	<link rel="stylesheet" href="/static/excel-bootstrap-table-filter-style.css">
+	<link rel="stylesheet" href="/onto-spread-ed/staticbootstrap-table-filter-style.css">
 
 	<!-- //fontawesome: -->
 	<link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.15.0/css/all.css" crossorigin="anonymous">
