+++ onto_spread_ed/utils/github.py
@@ -1,15 +1,21 @@
+from __future__ import annotations
+
 import base64
 import csv
 import io
 import logging
-from typing import List, Tuple, Dict
+import re
+from functools import cache
+from typing import List, Tuple, Dict, Optional
 
 import openpyxl
 from flask_github import GitHub
 
 _logger = logging.getLogger(__name__)
 
-def get_csv(github: GitHub, repository_name: str, folder: str, spreadsheet_name: str) -> Tuple[str, List[List[str]], List[str]]:
+
+def get_csv(github: GitHub, repository_name: str, folder: str, spreadsheet_name: str) -> Tuple[
+    str, List[List[str]], List[str]]:
     csv_file = github.get(
         f'repos/{repository_name}/contents/{folder}/{spreadsheet_name}'
     )
@@ -53,3 +59,16 @@ def get_spreadsheet(github: GitHub,
         _logger.error(f"Could not index {spreadsheet}: {e}")
 
     return file_sha, rows, header
+
+
+def get_spreadsheets(github: GitHub,
+                     repository_name: str,
+                     folder: str = "",
+                     exclude_pattern: Optional[re.Pattern | str] = None) -> list[str]:
+    tree = github.get(
+        f'repos/{repository_name}/git/trees/master',
+        params={"recursive": "true"}
+    )
+    entries = tree["tree"]
+
+    return [x["path"] for x in entries if x["path"].endswith(".xlsx") and not (exclude_pattern and re.match(exclude_pattern, x["path"]))]
